requirements.txt
fastmcp>=0.4.0
markdownify>=0.12.0
mcp>=1.0.0
pydantic>=2.5.0
readabilipy>=0.2.0
python-dotenv>=1.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
yt-dlp>=2024.1.0
openai>=0.27.0
open-meteo>=0.3.2
timezonefinder>=6.2.0
pytz>=2024.1
python-dateutil>=2.8.0
arxiv>=2.2.0
psycopg2-binary==2.9.7
sentence-transformers==2.7.0
transformers==4.36.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
huggingface-hub==0.19.4
flask==2.3.3
